+++ test/cases/loader-option-preprocessor-handlebars/webpack.config.js
@@ -1,0 +1,41 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+const Handlebars = require('handlebars');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  entry: {
+    index: './src/home.html',
+  },
+
+  plugins: [new HtmlBundlerPlugin()],
+
+  module: {
+    rules: [
+      {
+        test: /\.html$/,
+        loader: HtmlBundlerPlugin.loader,
+        options: {
+          preprocessor: (content, loaderContext) =>
+            Handlebars.compile(content)({
+              title: 'Breaking Bad',
+              firstname: 'Walter',
+              lastname: 'Heisenberg',
+            }),
+        },
+      },
+      {
+        test: /\.(png|svg|jpe?g|webp)$/i,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};